+++ custom/plugins/SwagCmsExtensions/src/Resources/app/storefront/src/swag-cms-extensions-scroll-navigation/swag-cms-extensions-scroll-navigation.plugin.js
@@ -94,7 +94,9 @@ export default class SwagCmsExtensionsScrollNavigation extends Plugin {
         this.anchoredSections = anchoredSections;
         this.navigationList = DomAccess.querySelector(this.sidebar, this.options.navigationListSelector);
 
-        this.enableSmoothScrolling(this.navigationList);
+        if (this.options.pageSettings.active === true) {
+            this.enableSmoothScrolling(this.navigationList);
+        }
         this.registerObserver(anchoredSections);
     }